Refactor and split my-sites/plans/jetpack-plans (utils and constants)
