+++ client/my-sites/plans/jetpack-plans/test/utils.js
@@ -16,7 +16,7 @@ import {
 	PLAN_JETPACK_SECURITY_REALTIME_MONTHLY,
 } from '@automattic/calypso-products';
 import { getProductCost } from 'calypso/state/products-list/selectors/get-product-cost';
-import { getHighestAnnualDiscount } from 'calypso/my-sites/plans/jetpack-plans/utils';
+import { getHighestAnnualDiscount } from 'calypso/my-sites/plans/jetpack-plans/utils/ui';
 
 let mockProductPrices;
